Add webmanifest and basic PWA support
For new Rails apps, a suitable webmanifest is added by default
