+++ config/initializers/mime_types.rb
@@ -4,3 +4,6 @@
 
 # Add new mime types for use in respond_to blocks:
 # Mime::Type.register "text/richtext", :rtf
+
+# https://w3c.github.io/manifest/#media-type-registration
+Mime::Type.register 'application/manifest+json', :webmanifest
